requirements.txt
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.166.0
gspread==6.1.4
pandas==2.2.3
python-dotenv==1.1.0
